:recycle: refactor: remove `useResponsive` hook
